v1.0.5.4. Remove more NULL check. Add NONBUS_DEVICE class. Change float to double. Modify logic of some BLOCKs. Remove some ALWAYS true check in sparse matrix. Performance almost unchanged.

+++ code/header/basic/area.h
@@ -42,7 +42,7 @@ class AREA
         unsigned int area_number;
         string area_name;
         unsigned int area_swing_bus;
-        float expected_power_leaving_area_in_MW;
-        float area_power_mismatch_tolerance_in_MW;
+        double expected_power_leaving_area_in_MW;
+        double area_power_mismatch_tolerance_in_MW;
 };
 #endif // AREA_H
